% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumseq.R
\name{cumseq}
\alias{cumseq}
\title{Create a sequence cumulatively}
\usage{
cumseq(x, fill_missing = NULL)
}
\arguments{
\item{x}{A vector}

\item{fill_missing}{A value to replace missing values}
}
\value{

}
\description{
Create an incremental sequence of numbers where there is a change
in the value of a vector.
}
\examples{
cumseq(c(0, 0, 1, 0, 1, 1, 0))
cumseq(c(0, 0, 1, 0, NA, NA, 0), fill_missing = 9)

}
